fix : 모든 파일의 구조, 폴더명 코드 수정
