+++ src/App.js
@@ -1,7 +1,6 @@
-import React from 'react';
-import { BrowserRouter } from 'react-router-dom';
-import Routers from './Component/router';
-import './App.css';
+import React from "react";
+import { BrowserRouter } from "react-router-dom";
+import Routers from "./components/router";
 
 function App() {
   return (
