--- src/Component/Announcement/an.jsx
@@ -1,12 +1,0 @@
-import React, { useState } from 'react';
-import * as S from "./style1";
-
-const App = () => {
-  return (
-    <S.ProjectContainer id="공지사항">
-        <S.Projects>DevCoop 공지사항</S.Projects>
-    </S.ProjectContainer>
-  );
-};
-
-export default App;
