+++ src/common/GlobalStyle.jsx
@@ -2,7 +2,7 @@ import { createGlobalStyle } from "styled-components";
 
 export const GlobalStyle = createGlobalStyle`
   body {
-    backgroundcolor: #000
+    background-color: #000
   }
 
   li,ol {
