+++ src/components/Mainpage/style.js
