+++ src/components/Notice/index.jsx
@@ -1,0 +1,12 @@
+import React from "react";
+import * as S from "./style";
+
+const Notice = () => {
+  return (
+    <S.ProjectContainer id="공지사항">
+      <S.Projects>DevCoop 공지사항</S.Projects>
+    </S.ProjectContainer>
+  );
+};
+
+export default Notice;
