+++ src/components/Notice/style.js
