+++ src/components/Team/style.js
