+++ src/pages/Mainpage/Mainpage.jsx
@@ -1,5 +1,5 @@
-import React from 'react';
-import Main from '../../Component/Mainpage/Mainpg';
+import React from "react";
+import Main from "../../components/Mainpage/Mainpg";
 
 export default function Mainpage() {
   return (
