+++ src/pages/Notfound.jsx
@@ -1,11 +1,7 @@
-import React from 'react';
+import React from "react";
 
 const Notfound = () => {
-    return (
-        <div>
-            404 Error
-        </div>
-    );
+  return <div>404 Error</div>;
 };
-  
+
 export default Notfound;
