+++ src/pages/Sign/Sign.jsx
@@ -1,10 +1,10 @@
-import React from 'react'
-import Sign from '../../Component/Sign/Signpage'
+import React from "react";
+import Sign from "../../components/Sign/Signpage";
 
 export default function Mainpage() {
   return (
     <>
-    <Sign/>
+      <Sign />
     </>
-  )
+  );
 }
